requirements.txt

torch>=2.0.0
torchvision>=0.15.0

diffusers>=0.34.0
accelerate>=1.0.0

timm==0.5.4
kornia>=0.7.0
lpips>=0.1.4

pillow>=10.0.0
opencv-python>=4.8.0

numpy>=1.24.0
scipy>=1.10.0

tqdm>=4.65.0
tensorboard>=2.14.0
info-nce-pytorch>=0.1.4

pyyaml>=6.0

